Move the delegate implementation back to its original file; adapt views to maximize shared code in abstract course list view
